Begin working on fettuccine
Fettuccine is the upgraded version of spaghetti that I am going to use
for LD34.

Right now it certainly is better - it is much cleaner, slightly better
documented, and I think that it is easier to use.

Right now the main new functionality is in the Renderer class, which
handles most basic types of 2D rendering. It can apply different matrix
transforms to images that it is drawing.

There is also now the KeyControl class, which can be used to track
whether certain keys are pressed or not. KeyControls are attached to the
Game objected through the watchControl() method.

I am working on making a Sprite system.